+++ NEWS.md
@@ -8,6 +8,9 @@
 - Data available for `upset_set_size()` now includes all metadata of the original data frame, enabling to annotate the bars, e.g.:
    - `upset_set_size(geom=geom_bar(aes(fill=mpaa, x=group)))`
 - Intersections can now be sorted by multiple criteria, e.g. first by degree and then by cardinality: `sort_intersections_by=c('degree', 'cardinality')` (#47)
+- Breaking: `dot_size` argument was removed. Use `matrix=intersection_matrix(geom=geom_point(size=5))` instead
+- Intersection matrix can now be customized, including the points (e.g. changing shape to squares), segments (e.g. using dotted line) and outlines (changing color) allowing to create [example 5.4](https://krassowski.github.io/complex-upset/articles/Examples_R.html#5-4-adjusting-the-intersection-matrix):
+ > ![](https://raw.githubusercontent.com/krassowski/complex-upset/master/tests/figs/examples/example-5-4-adjusting-the-intersection-matrix-1.svg)
 
 # Version 0.7.4